cli: Code fixes from the refactoring
